add light jets weights and njets arg
Initially, all jets were being used for the calculation of the weights. A new parameter ("njets") was added to define the number of jets per event that will be used for the weight calculation. Additionally, the "flavor" parameter was added to the "add_btag_weights()" method, allowing for the inclusion of b-tagging weights associated with both bc and light jets.